Updated Naming Convention and Attack
