--- unity-project/mini-game-collection/Assets/2024/Team11/Assets/PP_Free_Swords.fbx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:234c8e370cef12dd6af058981b9a665ca5d873e18ad3b3eeacd71661dc8337d2
-size 336972
