+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Prefab/2024_Team11_KyleRobot_Var_Green.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 5bfb8e9079c498743aadb2c4945af2c9
-folderAsset: yes
-DefaultImporter:
+guid: f42ebec61d4009d48ab3b153b15de376
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
