+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Prefab/2024_Team11_KyleRobot_Var_Purple.prefab
@@ -1,0 +1,222 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &2678896370019372531
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 7635806274571501702}
+  m_Layer: 0
+  m_Name: Player1_Spawner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &7635806274571501702
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2678896370019372531}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 1, z: 0, w: 0}
+  m_LocalPosition: {x: 0.021, y: 1.3639984, z: 0.765}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 7454348184364338473}
+  m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
+--- !u!1001 &6987778271729429186
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -6.98
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 31.238178
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 75.7
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 90
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_Name
+      value: KyleRobotTeam11Purple
+      objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5866666021909216657, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_Controller
+      value: 
+      objectReference: {fileID: 9100000, guid: cfb6d5e21d7412e4db8b01fc576d4bd3, type: 2}
+    - target: {fileID: 5866666021909216657, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_ApplyRootMotion
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7163341763079882450, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: d1fb8a66685070840914f6e30cc1b4a0, type: 2}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects:
+    - targetCorrespondingSourceObject: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      insertIndex: 1
+      addedObject: {fileID: 7635806274571501702}
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 213494343204747037}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 900123836727292387}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 8423247624700841032}
+  m_SourcePrefab: {fileID: 100100000, guid: 0968fe2872e11854eb25568199d0ffce, type: 3}
+--- !u!4 &7454348184364338473 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: -8679921383154817045, guid: 0968fe2872e11854eb25568199d0ffce,
+    type: 3}
+  m_PrefabInstance: {fileID: 6987778271729429186}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1 &7798258188153591699 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 0968fe2872e11854eb25568199d0ffce,
+    type: 3}
+  m_PrefabInstance: {fileID: 6987778271729429186}
+  m_PrefabAsset: {fileID: 0}
+--- !u!54 &213494343204747037
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7798258188153591699}
+  serializedVersion: 4
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_CenterOfMass: {x: 0, y: 0, z: 0}
+  m_InertiaTensor: {x: 1, y: 1, z: 1}
+  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ImplicitCom: 1
+  m_ImplicitTensor: 1
+  m_UseGravity: 1
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 80
+  m_CollisionDetection: 0
+--- !u!114 &900123836727292387
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7798258188153591699}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8a4b37f9a19991943901669478a855f0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  <PlayerID>k__BackingField: 1
+  moveSpeed: 2
+  rotationSpeed: 4
+  isJumping: 0
+  isGrounded: 0
+  isAttacking: 0
+  respawnPoint: {fileID: 0}
+  spherePrefab: {fileID: 5428847396346318475, guid: e553641db2a95764b8753912a6412d5d,
+    type: 3}
+  spawningPos: {fileID: 7635806274571501702}
+  sphereSpeed: 10
+--- !u!136 &8423247624700841032
+CapsuleCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7798258188153591699}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Radius: 0.5
+  m_Height: 1.843324
+  m_Direction: 1
+  m_Center: {x: 0, y: 0.8774675, z: 0}
