+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Scripts/T11_AttackBall.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class T11_AttackBall : MonoBehaviour
+{
+    public float life = 3;
+
+    private void Awake()
+    {
+        Destroy(gameObject, life);
+    }
+
+    private void OnCollisionEnter(Collision collision)
+    {
+        if (collision.gameObject.CompareTag("Collider"))
+        {
+            Destroy(gameObject);
+        }
+    }
+}
