+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Scripts/T11_AttackBall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aee5e35d70fa46e45a56b2a8904cc56a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
