+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Texture and Materials/2024-team11-color-dark_red.mat
@@ -7,7 +7,7 @@ Material:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Arena2
+  m_Name: 2024-team11-color-dark_red
   m_Shader: {fileID: 4800000, guid: 933532a4fcc9baf4fa0491de14d08ed7, type: 3}
   m_Parent: {fileID: 0}
   m_ModifiedSerializedProperties: 0
@@ -113,12 +113,12 @@ Material:
     - _WorkflowMode: 1
     - _ZWrite: 1
     m_Colors:
-    - _BaseColor: {r: 0.6886792, g: 0.46248263, b: 0.061721236, a: 1}
-    - _Color: {r: 0.6886792, g: 0.46248257, b: 0.061721213, a: 1}
+    - _BaseColor: {r: 0.3301887, g: 0, b: 0, a: 1}
+    - _Color: {r: 0.33018863, g: 0, b: 0, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 1}
   m_BuildTextureStacks: []
---- !u!114 &1670643583444031879
+--- !u!114 &759754333596765554
 MonoBehaviour:
   m_ObjectHideFlags: 11
   m_CorrespondingSourceObject: {fileID: 0}
