+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Texture and Materials/2024-team11-color-green.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da4981c2d081b9a449cfbc7194e9fcc8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
