+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Texture and Materials/2024-team11-color-purple.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b0d6ebbf945003469db65bbbbac4151
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
