+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Texture and Materials/T11_Robot_Textures/2024-team11-Robot_Albedo_Green_Var.Tex.png.meta
