+++ unity-project/mini-game-collection/Assets/2024/Team11/T11_Texture and Materials/T11_Robot_Textures/KyleRobot_Metallic - Copy.png.meta
