+++ unity-project/mini-game-collection/ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - Arena
+  - Collider
   layers:
   - Default
   - TransparentFX
